+++ src/components/examList/ExamList.tsx
@@ -1,6 +1,8 @@
-import { useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
 import ExamCard from './ExamCard'
-import type { TabId, Tab, ExamData } from '@/types/examList'
+import type { TabId, Tab } from '@/types/examList'
+import { infiniteScrolling } from '@/hooks/examList/useInfiniteScroll'
+import { useInfiniteQuery } from '@tanstack/react-query'
 
 const TABS: Tab[] = [
   { id: 'all', label: '전체보기' },
@@ -8,197 +10,51 @@ const TABS: Tab[] = [
   { id: 'notStart', label: '미응시' },
 ]
 
-const mockExamData: ExamData[] = [
-  // 프론트엔드 과목
-  {
-    id: 'html-css',
-    title: 'HTML 기초',
-    category: 'frontend',
-    technology: 'HTML/CSS',
-    description: 'HTML · 80점/100점 · 8/10개 정답',
-    status: 'completed',
-    score: 80,
-    totalScore: 100,
-    correctAnswers: 8,
-    totalQuestions: 10,
-    icon: '/examList/html.png',
-  },
-  {
-    id: 'javascript-basic',
-    title: 'JavaScript 기초',
-    category: 'frontend',
-    technology: 'JavaScript',
-    description: 'JavaScript · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/nodejs.png',
-  },
-  {
-    id: 'github-usage',
-    title: 'Github 응용',
-    category: 'frontend',
-    technology: 'Github',
-    description: 'github · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/github.png',
-  },
-  {
-    id: 'react-basic',
-    title: 'React 실습',
-    category: 'frontend',
-    technology: 'React',
-    description: 'React · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/Frame.png',
-  },
-  {
-    id: 'nodejs-basic',
-    title: 'Node.js 기초',
-    category: 'frontend',
-    technology: 'Node.js',
-    description: 'Node.js · 90점/100점 · 9/10개 정답',
-    status: 'completed',
-    score: 90,
-    totalScore: 100,
-    correctAnswers: 9,
-    totalQuestions: 10,
-    icon: '/examList/nodejs.png',
-  },
-  {
-    id: 'database-basic',
-    title: 'Database 기초',
-    category: 'frontend',
-    technology: 'Database',
-    description: 'Database · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/amazon-database.png',
-  },
-  {
-    id: 'typescript-basic',
-    title: 'TypeScript 심화',
-    category: 'frontend',
-    technology: 'TypeScript',
-    description: 'TypeScript · 75점/100점 · 7.5/10개 정답',
-    status: 'completed',
-    score: 75,
-    totalScore: 100,
-    correctAnswers: 7.5,
-    totalQuestions: 10,
-    icon: '/examList/typescript.png',
-  },
-  {
-    id: 'aws-frontend',
-    title: 'AWS 심화',
-    category: 'frontend',
-    technology: 'AWS',
-    description: 'AWS · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/aws.png',
-  },
-  {
-    id: 'react-native',
-    title: 'React Native 기초',
-    category: 'frontend',
-    technology: 'React Native',
-    description: 'React Native · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/Frame.png',
-  },
-  // 백엔드 과목
-  {
-    id: 'python-basic',
-    title: 'Python 기초',
-    category: 'backend',
-    technology: 'Python',
-    description: 'Python · 85점/100점 · 8.5/10개 정답',
-    status: 'completed',
-    score: 85,
-    totalScore: 100,
-    correctAnswers: 8.5,
-    totalQuestions: 10,
-    icon: '/examList/python.png',
-  },
-  {
-    id: 'html-css-backend',
-    title: 'HTML/CSS 실습',
-    category: 'backend',
-    technology: 'HTML/CSS',
-    description: 'HTML/CSS · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/html.png',
-  },
-  {
-    id: 'javascript-backend',
-    title: 'JavaScript 심화',
-    category: 'backend',
-    technology: 'JavaScript',
-    description: 'JavaScript · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/nodejs.png',
-  },
-  {
-    id: 'database-backend',
-    title: 'Database 심화',
-    category: 'backend',
-    technology: 'Database',
-    description: 'Database · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/amazon-database.png',
-  },
-  {
-    id: 'django-basic',
-    title: 'Django 기초',
-    category: 'backend',
-    technology: 'Django',
-    description: 'Django · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/django.png',
-  },
-  {
-    id: 'fastapi-basic',
-    title: 'FastAPI 실습',
-    category: 'backend',
-    technology: 'FastAPI',
-    description: 'FastAPI · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: '/examList/fastapi.png',
-  },
-  {
-    id: 'flask-basic',
-    title: 'Flask 기초',
-    category: 'backend',
-    technology: 'Flask',
-    description: 'Flask · 응시하고 점수를 확인해보세요!',
-    status: 'notStart',
-    icon: 'flask',
-  },
-]
-
 export default function ExamList() {
   const [activeTab, setActiveTab] = useState<TabId>('all')
+  const observerRef = useRef<HTMLDivElement>(null)
 
-  const filteredExams = mockExamData.filter((exam) => {
-    switch (activeTab) {
-      case 'all':
-        return true
-      case 'completed':
-        return exam.status === 'completed'
-      case 'notStart':
-        return exam.status === 'notStart'
-      default:
-        return true
-    }
+  const {
+    data,
+    isLoading,
+    isError,
+    isFetchingNextPage,
+    hasNextPage,
+    fetchNextPage,
+  } = useInfiniteQuery({
+    queryKey: ['exam', activeTab],
+    queryFn: ({ pageParam = 1 }) => infiniteScrolling(pageParam, activeTab, 5),
+    getNextPageParam: (lastPage) => lastPage.nextPage,
+    initialPageParam: 1,
   })
 
+  const allExams = data?.pages.flatMap((p) => p.data) ?? []
+
+  useEffect(() => {
+    const observer = new IntersectionObserver(
+      (entries) => {
+        if (entries[0].isIntersecting && hasNextPage && !isFetchingNextPage) {
+          fetchNextPage()
+        }
+      },
+      { threshold: 0.1 }
+    )
+    if (observerRef.current) {
+      observer.observe(observerRef.current)
+    }
+    return () => observer.disconnect()
+  }, [data, isFetchingNextPage, hasNextPage])
+
   const handleTakeExam = (examId: string): void => {
-    // TODO: 실제 시험 페이지로 라우팅
     console.log('응시하기:', examId)
   }
 
   const handleViewDetails = (examId: string): void => {
-    // TODO: 시험 결과 상세 페이지로 라우팅
     console.log('상세보기:', examId)
   }
 
+  if (isLoading) return <p>Loading...</p>
+  if (isError) return <p>Error..</p>
   return (
     <div className="flex-1">
       <h2 className="text-[32px] font-bold mb-6">쪽지시험</h2>
@@ -224,21 +80,42 @@ export default function ExamList() {
 
       {/* 시험 카드 리스트 */}
       <div className="space-y-4">
-        {filteredExams.length > 0 ? (
-          filteredExams.map((exam) => (
-            <ExamCard
-              key={exam.id}
-              exam={exam}
-              onTakeExam={handleTakeExam}
-              onViewDetails={handleViewDetails}
-            />
-          ))
+        {allExams.length > 0 ? (
+          <>
+            {allExams.map((exam) => (
+              <ExamCard
+                key={exam.id}
+                exam={exam}
+                onTakeExam={handleTakeExam}
+                onViewDetails={handleViewDetails}
+              />
+            ))}
+
+            {/* 로딩 인디케이터 */}
+            {isFetchingNextPage && (
+              <div className="text-center py-8">
+                <div className="inline-block animate-spin rounded-full h-8 w-8 border-b-2 border-purple-600"></div>
+                <p className="mt-2 text-gray-500">
+                  더 많은 시험을 불러오는 중...
+                </p>
+              </div>
+            )}
+
+            {/* 마지막 페이지 표시 */}
+            {!hasNextPage && allExams.length > 0 && (
+              <div className="text-center py-8">
+                <p className="text-gray-500">모든 시험을 불러왔습니다.</p>
+              </div>
+            )}
+          </>
         ) : (
           <div className="text-center py-12">
             <p className="text-gray-500 text-lg">해당하는 시험이 없습니다.</p>
           </div>
         )}
       </div>
+
+      <div ref={observerRef} className="h-4" />
     </div>
   )
 }
